requirements.txt
aniso8601==7.0.0
asgiref==3.2.5
astroid==2.3.3
attrs==19.3.0
dj-database-url==0.5.0
Django==3.0.4
django-filter==2.2.0
django-heroku==0.3.1
entrypoints==0.3
flake8==3.7.9
graphene==2.1.8
graphene-django==2.8.2
graphql-core==2.3.1
graphql-relay==2.0.1
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
promise==2.3
py==1.8.1
pycodestyle==2.5.0
psycopg2
mysqlclient==1.4.6
pyflakes==2.1.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.4.1
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
wcwidth==0.1.8
whitenoise==5.0.1
wrapt==1.11.2
